refactor-choose-shose-상태관리---jobeomjun--1차 리팩토리
