+++ src/components/choose-shose/text-search/mainContainder/maincontainer.css.ts
@@ -4,7 +4,7 @@ import { theme } from '../../../../styles/theme';
 export const MainContainter_Background = style({
   display: 'flex',
   justifyContent: 'center',
-  marginTop: '155px',
+  marginTop: '65px',
   marginBottom: '108px',
 });
